Homepage Redesign
At the top we have this list of featured apps, this is good to
give emphasis on some specific applications.

Then we have some category areas displaying up to 9 apps, if the
window is big enough it'll show an grid otherwise it'll show a list.

These changes are great because they increase the card density on the
desktop and differentiate the featured apps from the top rated apps
from each category.

Signed-off-by: Felipe Kinoshita <kinofhek@gmail.com>
Co-authored-by: Carl Schwan <carl@carlschwan.eu>